move market highlight to a new file
